feat: Add user authentication

+++ app/Http/Controllers/ChecklistController.php
@@ -9,6 +9,11 @@ use Illuminate\Support\Facades\Auth;
 
 class ChecklistController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index','show']]);
+    }
+
     public function index()
     {
         $checklists = Checklist::simplePaginate(10);
@@ -27,9 +32,9 @@ class ChecklistController extends Controller
     {
         $checklist = Checklist::findOrFail($id);
 
-        $auth = Auth::user();
+        $this->authorize('edit', $checklist);
 
-        return view('checklist/edit', compact('checklist', 'auth'));
+        return view('checklist/edit', compact('checklist'));
     }
 
     public function update(Request $request, $id)
@@ -41,6 +46,9 @@ class ChecklistController extends Controller
         ]);
 
         $checklist = Checklist::findOrFail($id);
+
+        $this->authorize('edit', $checklist);
+
         $checklist->checks = $request->checks;
         $checklist->description = $request->description;
         $checklist->name = $request->name;
@@ -52,6 +60,9 @@ class ChecklistController extends Controller
     public function destroy($id)
     {
         $checklist = Checklist::findOrFail($id);
+
+        $this->authorize('edit', $checklist);
+
         $checklist->delete();
 
         return redirect("/checklist");
